requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
coloredlogs==15.0.1
flatbuffers==23.5.26
humanfriendly==10.0
idna==3.6
imageio==2.33.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
lazy_loader==0.3
llvmlite==0.41.1
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
onnxruntime==1.16.3
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.1.0
platformdirs==4.1.0
pooch==1.8.0
protobuf==4.25.1
PyMatting==1.1.12
python-dotenv==1.0.0
referencing==0.32.0
rembg==2.0.53
requests==2.31.0
rpds-py==0.13.2
scikit-image==0.22.0
scipy==1.11.4
soupsieve==2.5
sympy==1.12
tifffile==2023.12.9
tqdm==4.66.1
urllib3==2.1.0
